Added runtime loading of assemblies
