+++ InterfacesAndAbstractions/InterfacesAndAbstractions.sln
@@ -2,6 +2,8 @@
 Microsoft Visual Studio Solution File, Format Version 12.00
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ShapePrinter", "ShapePrinter\ShapePrinter.csproj", "{9F2E87E3-EBA8-4A74-8E58-C2FBA66B12CD}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ExtensionLibrary", "ExtensionLibrary\ExtensionLibrary.csproj", "{194FE04E-E8E4-461E-A6A3-B456D4E9BAF5}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -12,5 +14,9 @@ Global
 		{9F2E87E3-EBA8-4A74-8E58-C2FBA66B12CD}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{9F2E87E3-EBA8-4A74-8E58-C2FBA66B12CD}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{9F2E87E3-EBA8-4A74-8E58-C2FBA66B12CD}.Release|Any CPU.Build.0 = Release|Any CPU
+		{194FE04E-E8E4-461E-A6A3-B456D4E9BAF5}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{194FE04E-E8E4-461E-A6A3-B456D4E9BAF5}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{194FE04E-E8E4-461E-A6A3-B456D4E9BAF5}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{194FE04E-E8E4-461E-A6A3-B456D4E9BAF5}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 EndGlobal
