+++ InterfacesAndAbstractions/ShapePrinter/UiHandler.cs
@@ -1,6 +1,8 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.IO;
 using System.Linq;
+using System.Reflection;
 
 namespace ShapePrinter
 {
@@ -8,9 +10,7 @@ namespace ShapePrinter
     {
         private static Type _type = typeof(IPrintable);
 
-        private static List<Type> _types = AppDomain.CurrentDomain.GetAssemblies()
-            .SelectMany(s => s.GetTypes())
-            .Where(p => _type.IsAssignableFrom(p) && !p.IsInterface && !p.IsAbstract).ToList();
+        private static List<Type> _types = GetAssemblyTypes();
 
         public static void Run()
         {
@@ -27,7 +27,7 @@ namespace ShapePrinter
                     {
                         var text = GetText();
                         objToPrint =
-                            (IPrintable)Activator.CreateInstance(printableType, new object[] { text});
+                            (IPrintable)Activator.CreateInstance(printableType, new object[] { text });
                     }
                     else
                     {
@@ -186,6 +186,34 @@ namespace ShapePrinter
         {
             Console.WriteLine("\nDraw a new picture? 'Y' to continue");
             return Console.ReadKey().Key is ConsoleKey.Y;
+        }
+
+        private static List<Type> GetAssemblyTypes()
+        {
+            string currentAssemblyFolderPath = AppDomain.CurrentDomain.BaseDirectory;
+            string[] filePaths = Directory.GetFiles(currentAssemblyFolderPath, "*.dll",
+                SearchOption.TopDirectoryOnly);
+
+            List<string> UsedAssebmliesFilePaths = new List<string>();
+
+            foreach (var assembly in AppDomain.CurrentDomain.GetAssemblies())
+            {
+                UsedAssebmliesFilePaths.Add(assembly.Location);
+            }
+
+
+            foreach (string filePath in filePaths)
+            {
+                if (!UsedAssebmliesFilePaths.Contains(filePath))
+                {
+                    Assembly assembly = Assembly.LoadFrom(filePath);
+                    AppDomain.CurrentDomain.Load(assembly.GetName());
+                }
+            }
+
+            return AppDomain.CurrentDomain.GetAssemblies()
+                .SelectMany(s => s.GetTypes())
+                .Where(p => _type.IsAssignableFrom(p) && !p.IsInterface && !p.IsAbstract).ToList();
         }
     }
 }
